Create OnboardingFlowListener
