+++ app/src/main/java/com/tari/android/wallet/ui/fragment/onboarding/activity/OnboardingFlowListener.kt
@@ -1,0 +1,8 @@
+package com.tari.android.wallet.ui.fragment.onboarding.activity
+
+interface OnboardingFlowListener {
+    fun continueToEnableAuth()
+    fun continueToCreateWallet()
+    fun onAuthSuccess()
+    fun navigateToNetworkSelection()
+}
