--- app/src/main/java/com/tari/android/wallet/ui/fragment/onboarding/createWallet/CreateWalletListener.kt
@@ -1,5 +1,0 @@
-package com.tari.android.wallet.ui.fragment.onboarding.createWallet
-
-interface CreateWalletListener {
-    fun continueToEnableAuth()
-}
