+++ app/src/main/java/com/tari/android/wallet/ui/fragment/onboarding/inroduction/IntroductionFragment.kt
@@ -36,7 +36,6 @@ import android.animation.AnimatorSet
 import android.animation.ObjectAnimator
 import android.animation.ValueAnimator
 import android.media.AudioManager
-import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
@@ -51,12 +50,27 @@ import androidx.fragment.app.viewModels
 import com.daasuu.ei.Ease
 import com.daasuu.ei.EasingInterpolator
 import com.tari.android.wallet.R
-import com.tari.android.wallet.R.string.*
+import com.tari.android.wallet.R.string.create_wallet_privacy_policy
+import com.tari.android.wallet.R.string.create_wallet_user_agreement
+import com.tari.android.wallet.R.string.create_wallet_user_agreement_and_privacy_policy
+import com.tari.android.wallet.R.string.introduction_selected_wallet
+import com.tari.android.wallet.R.string.privacy_policy_url
+import com.tari.android.wallet.R.string.user_agreement_url
 import com.tari.android.wallet.databinding.FragmentIntroductionBinding
 import com.tari.android.wallet.extension.applyURLStyle
-import com.tari.android.wallet.ui.common.CommonFragment
-import com.tari.android.wallet.ui.extension.*
-import com.tari.android.wallet.ui.fragment.onboarding.activity.OnboardingFlowActivity
+import com.tari.android.wallet.ui.extension.addAnimatorListener
+import com.tari.android.wallet.ui.extension.animateClick
+import com.tari.android.wallet.ui.extension.doOnGlobalLayout
+import com.tari.android.wallet.ui.extension.getResourceUri
+import com.tari.android.wallet.ui.extension.gone
+import com.tari.android.wallet.ui.extension.invisible
+import com.tari.android.wallet.ui.extension.postDelayed
+import com.tari.android.wallet.ui.extension.setLayoutSize
+import com.tari.android.wallet.ui.extension.setOnThrottledClickListener
+import com.tari.android.wallet.ui.extension.string
+import com.tari.android.wallet.ui.extension.temporarilyDisableClick
+import com.tari.android.wallet.ui.extension.visible
+import com.tari.android.wallet.ui.fragment.onboarding.activity.OnboardingFlowFragment
 import com.tari.android.wallet.ui.fragment.restore.activity.WalletRestoreActivity
 import com.tari.android.wallet.ui.fragment.settings.allSettings.TariVersionModel
 import com.tari.android.wallet.util.Constants
@@ -70,7 +84,7 @@ import kotlin.math.min
  * @author The Tari Development Team
  */
 
-class IntroductionFragment : CommonFragment<FragmentIntroductionBinding, IntroductionViewModel>() {
+class IntroductionFragment : OnboardingFlowFragment<FragmentIntroductionBinding, IntroductionViewModel>() {
 
     private val handler = Handler(Looper.getMainLooper())
 
@@ -147,7 +161,7 @@ class IntroductionFragment : CommonFragment<FragmentIntroductionBinding, Introdu
                 setupAndStartVideo()
             }
             createWalletButton.setOnThrottledClickListener { onCreateWalletClick() }
-            selectNetworkContainerView.setOnThrottledClickListener { (requireActivity() as OnboardingFlowActivity).navigateToNetworkSelection() }
+            selectNetworkContainerView.setOnThrottledClickListener { onboardingListener.navigateToNetworkSelection() }
         }
     }
 
@@ -155,9 +169,7 @@ class IntroductionFragment : CommonFragment<FragmentIntroductionBinding, Introdu
         val size = min(ui.videoOuterContainerView.width, ui.videoOuterContainerView.height)
         ui.videoInnerContainerView.setLayoutSize(size, size)
         ui.rainAnimationVideoView.setVideoURI(requireContext().getResourceUri(R.raw.purple_orb))
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            ui.rainAnimationVideoView.setAudioFocusRequest(AudioManager.AUDIOFOCUS_NONE)
-        }
+        ui.rainAnimationVideoView.setAudioFocusRequest(AudioManager.AUDIOFOCUS_NONE)
         startVideo()
         videoViewHasBeenSetup = true
     }
@@ -228,7 +240,7 @@ class IntroductionFragment : CommonFragment<FragmentIntroductionBinding, Introdu
             addListener(onEnd = { playTariWalletLottieAnimation() })
         }
 
-        ui.tariLogoLottieAnimationView.addAnimatorListener(onEnd = { (requireActivity() as? IntroductionListener)?.continueToCreateWallet() })
+        ui.tariLogoLottieAnimationView.addAnimatorListener(onEnd = { onboardingListener.continueToCreateWallet() })
 
         val tariViewScaleAnim = ValueAnimator.ofFloat(ui.tariLogoLottieAnimationView.scaleX, 1f).apply {
             duration = Constants.UI.CreateWallet.tariTextAnimViewDurationMs
