--- app/src/main/java/com/tari/android/wallet/ui/fragment/onboarding/inroduction/IntroductionListener.kt
@@ -1,5 +1,0 @@
-package com.tari.android.wallet.ui.fragment.onboarding.inroduction
-
-interface IntroductionListener {
-    fun continueToCreateWallet()
-}
