+++ app/src/main/java/com/tari/android/wallet/ui/fragment/onboarding/localAuth/LocalAuthModel.kt
@@ -1,12 +1,13 @@
 package com.tari.android.wallet.ui.fragment.onboarding.localAuth
 
 object LocalAuthModel {
+    data class SecureState(
+        val biometricsAvailable: Boolean = true,
+        val pinCodeSecured: Boolean = false,
+        val biometricsSecured: Boolean = false,
+    )
 
     sealed interface Effect {
         data object OnAuthSuccess : Effect
-    }
-
-    interface LocalAuthListener {
-        fun onAuthSuccess()
     }
 }
