--- app/src/main/java/com/tari/android/wallet/ui/fragment/onboarding/localAuth/SecureState.kt
@@ -1,7 +1,0 @@
-package com.tari.android.wallet.ui.fragment.onboarding.localAuth
-
-data class SecureState(
-    val biometricsAvailable: Boolean = true,
-    val pinCodeSecured: Boolean = false,
-    val biometricsSecured: Boolean = false,
-)
